flask-app/requirements.txt
anyio==3.7.0
appdirs==1.4.4
base58==2.1.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==3.0.10
fonttools==4.40.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
IPFS-Toolkit==0.5.13
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
multiaddr==0.0.9
netaddr==0.8.0
networkx==3.1
numpy==1.25.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
psutil==5.9.5
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
sympy==1.12
termcolor==2.3.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
ultralytics==8.0.122
urllib3==2.0.3
varint==1.0.2
Werkzeug==2.3.6
